+++ src/utils/FileSystem.ts
@@ -131,6 +131,7 @@ Type 'blog' to read posts.
     bin.children!.set('pwd', this.createFileNode('pwd', '[Core command: pwd]'));
     bin.children!.set('cat', this.createFileNode('cat', '[Core command: cat]'));
     bin.children!.set('tree', this.createFileNode('tree', '[Core command: tree]'));
+    bin.children!.set('render', this.createFileNode('render', '[Core command: render]'));
 
     // /usr/local/bin directory (custom/content commands)
     const local = this.createDirectoryNode('local');